+++ internal/storage/v2/grpc/tracereader.go
@@ -12,7 +12,6 @@ import (
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	"google.golang.org/grpc"
 
-	"github.com/jaegertracing/jaeger/internal/jpcommon"
 	"github.com/jaegertracing/jaeger/internal/storage/v2/api/tracestore"
 	"github.com/jaegertracing/jaeger/proto-gen/storage/v2"
 )
@@ -106,11 +105,83 @@ func toProtoQueryParameters(t tracestore.TraceQueryParams) *storage.TraceQueryPa
 	return &storage.TraceQueryParameters{
 		ServiceName:   t.ServiceName,
 		OperationName: t.OperationName,
-		Attributes:    jpcommon.ConvertMapToKeyValueList(t.Attributes),
+		Attributes:    convertMapToKeyValueList(t.Attributes),
 		StartTimeMin:  t.StartTimeMin,
 		StartTimeMax:  t.StartTimeMax,
 		DurationMin:   t.DurationMin,
 		DurationMax:   t.DurationMax,
 		SearchDepth:   int32(t.SearchDepth), //nolint: gosec // G115
+	}
+}
+
+func convertMapToKeyValueList(m pcommon.Map) []*storage.KeyValue {
+	keyValues := make([]*storage.KeyValue, 0, m.Len())
+	m.Range(func(k string, v pcommon.Value) bool {
+		keyValues = append(keyValues, &storage.KeyValue{
+			Key:   k,
+			Value: convertValueToAnyValue(v),
+		})
+		return true
+	})
+	return keyValues
+}
+
+func convertValueToAnyValue(v pcommon.Value) *storage.AnyValue {
+	switch v.Type() {
+	case pcommon.ValueTypeStr:
+		return &storage.AnyValue{
+			Value: &storage.AnyValue_StringValue{
+				StringValue: v.Str(),
+			},
+		}
+	case pcommon.ValueTypeBool:
+		return &storage.AnyValue{
+			Value: &storage.AnyValue_BoolValue{
+				BoolValue: v.Bool(),
+			},
+		}
+	case pcommon.ValueTypeInt:
+		return &storage.AnyValue{
+			Value: &storage.AnyValue_IntValue{
+				IntValue: v.Int(),
+			},
+		}
+	case pcommon.ValueTypeDouble:
+		return &storage.AnyValue{
+			Value: &storage.AnyValue_DoubleValue{
+				DoubleValue: v.Double(),
+			},
+		}
+	case pcommon.ValueTypeBytes:
+		return &storage.AnyValue{
+			Value: &storage.AnyValue_BytesValue{
+				BytesValue: v.Bytes().AsRaw(),
+			},
+		}
+	case pcommon.ValueTypeSlice:
+		arr := v.Slice()
+		arrayValues := make([]*storage.AnyValue, 0, arr.Len())
+		for i := 0; i < arr.Len(); i++ {
+			arrayValues = append(arrayValues, convertValueToAnyValue(arr.At(i)))
+		}
+		return &storage.AnyValue{
+			Value: &storage.AnyValue_ArrayValue{
+				ArrayValue: &storage.ArrayValue{
+					Values: arrayValues,
+				},
+			},
+		}
+	case pcommon.ValueTypeMap:
+		kvList := &storage.KeyValueList{}
+		v.Map().Range(func(k string, val pcommon.Value) bool {
+			kvList.Values = append(kvList.Values, &storage.KeyValue{
+				Key:   k,
+				Value: convertValueToAnyValue(val),
+			})
+			return true
+		})
+		return &storage.AnyValue{Value: &storage.AnyValue_KvlistValue{KvlistValue: kvList}}
+	default:
+		return nil
 	}
 }
