+++ internal/storage/v2/grpc/tracereader_test.go
@@ -301,3 +301,157 @@ func TestTraceReader_FindTraceIDs(t *testing.T) {
 		})
 	}
 }
+
+func TestConvertMapToKeyValueList(t *testing.T) {
+	tests := []struct {
+		name       string
+		attributes pcommon.Map
+		expected   []*storage.KeyValue
+	}{
+		{
+			name:       "empty map",
+			attributes: pcommon.NewMap(),
+			expected:   []*storage.KeyValue{},
+		},
+		{
+			name: "empty value",
+			attributes: func() pcommon.Map {
+				m := pcommon.NewMap()
+				m.PutEmpty("key1")
+				return m
+			}(),
+			expected: []*storage.KeyValue{
+				{
+					Key:   "key1",
+					Value: nil,
+				},
+			},
+		},
+		{
+			name: "primitive types",
+			attributes: func() pcommon.Map {
+				m := pcommon.NewMap()
+				m.PutStr("key1", "value1")
+				m.PutInt("key2", 42)
+				m.PutDouble("key3", 3.14)
+				m.PutBool("key4", true)
+				m.PutEmptyBytes("key5").Append(1, 2)
+				return m
+			}(),
+			expected: []*storage.KeyValue{
+				{
+					Key: "key1",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_StringValue{
+							StringValue: "value1",
+						},
+					},
+				},
+				{
+					Key: "key2",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_IntValue{
+							IntValue: 42,
+						},
+					},
+				},
+				{
+					Key: "key3",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_DoubleValue{
+							DoubleValue: 3.14,
+						},
+					},
+				},
+				{
+					Key: "key4",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_BoolValue{
+							BoolValue: true,
+						},
+					},
+				},
+				{
+					Key: "key5",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_BytesValue{
+							BytesValue: []byte{1, 2},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "nested map",
+			attributes: func() pcommon.Map {
+				m := pcommon.NewMap()
+				nested := pcommon.NewMap()
+				nested.PutStr("nestedKey", "nestedValue")
+				nested.CopyTo(m.PutEmptyMap("key1"))
+				return m
+			}(),
+			expected: []*storage.KeyValue{
+				{
+					Key: "key1",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_KvlistValue{
+							KvlistValue: &storage.KeyValueList{
+								Values: []*storage.KeyValue{
+									{
+										Key: "nestedKey",
+										Value: &storage.AnyValue{
+											Value: &storage.AnyValue_StringValue{
+												StringValue: "nestedValue",
+											},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "array attribute",
+			attributes: func() pcommon.Map {
+				m := pcommon.NewMap()
+				arr := pcommon.NewValueSlice()
+				arr.Slice().AppendEmpty().SetStr("value1")
+				arr.Slice().AppendEmpty().SetInt(42)
+				arr.Slice().CopyTo(m.PutEmptySlice("key1"))
+				return m
+			}(),
+			expected: []*storage.KeyValue{
+				{
+					Key: "key1",
+					Value: &storage.AnyValue{
+						Value: &storage.AnyValue_ArrayValue{
+							ArrayValue: &storage.ArrayValue{
+								Values: []*storage.AnyValue{
+									{
+										Value: &storage.AnyValue_StringValue{
+											StringValue: "value1",
+										},
+									},
+									{
+										Value: &storage.AnyValue_IntValue{
+											IntValue: 42,
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			result := convertMapToKeyValueList(test.attributes)
+			require.Equal(t, test.expected, result)
+		})
+	}
+}
